Move Identity+FinancialConnectionsSDK error logging code to AnalyticLoggableErrorV2
